+++ src/services/cart.services.js
@@ -1,0 +1,5 @@
+class CartServices {
+    async createCart() {
+        return await cartDao.create();
+    }
+}
